fix: resolve reported issues
fix Observer display issue
fix PlayWord embedding issue
improve HTML parsing method before embedding
optimize prompts
