+++ package.json
@@ -30,22 +30,21 @@
   "devDependencies": {
     "@commitlint/cli": "^19.7.1",
     "@commitlint/config-conventional": "^19.7.1",
-    "@eslint/js": "^9.20.0",
-    "@types/eslint__js": "^8.42.3",
-    "@types/node": "^22.13.4",
-    "@vitest/coverage-v8": "^3.0.6",
-    "eslint": "^9.20.1",
+    "@eslint/js": "^9.21.0",
+    "@types/node": "^22.13.9",
+    "@vitest/coverage-v8": "^3.0.7",
+    "eslint": "^9.21.0",
     "husky": "^9.1.7",
-    "lerna": "^8.2.0",
+    "lerna": "^8.2.1",
     "lint-staged": "^15.4.3",
-    "prettier": "^3.5.1",
-    "typescript": "^5.7.3",
-    "typescript-eslint": "^8.24.1",
-    "vite": "^6.1.1",
-    "vitest": "^3.0.6"
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.26.0",
+    "vite": "^6.2.0",
+    "vitest": "^3.0.7"
   },
   "optionalDependencies": {
-    "@rollup/rollup-linux-x64-gnu": "^4.34.8"
+    "@rollup/rollup-linux-x64-gnu": "^4.34.9"
   },
   "engines": {
     "node": ">=20"
@@ -61,6 +60,6 @@
     ]
   },
   "dependencies": {
-    "tsup": "8.3.6"
+    "tsup": "8.4.0"
   }
 }
