+++ packages/core/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './css'
 export * from './dom'
 export * from './html'
 export * from './logger'
