--- test/core/mocks/mockObserverUI.html
@@ -1,35 +1,0 @@
-<!-- Purpose: Mock Observer UI for testing dom manipulation -->
-<!doctype html>
-<html>
-  <head></head>
-  <body>
-    <div id="app"></div>
-    <div class="plwd-panel">
-      <div class="plwd-banner"></div>
-      <div class="plwd-input-box">
-        <input class="plwd-input" />
-        <div class="plwd-loader-box"></div>
-        <button class="plwd-cancel-btn"></button>
-        <button class="plwd-accept-btn"></button>
-      </div>
-      <div class="plwd-preview">
-        <p class="plwd-preview-title"></p>
-        <button class="plwd-dry-run-btn"></button>
-        <button class="plwd-clear-btn"></button>
-        <ul class="plwd-timeline">
-          <li class="plwd-timeline-item">
-            <span class="plwd-marker">✓</span>
-            input-value
-            <button class="plwd-delete-btn" data-index="0">
-              <svg class="plwd-trash-icon"></svg>
-            </button>
-          </li>
-        </ul>
-      </div>
-    </div>
-    <div class="plwd-toast">
-      <div class="plwd-toast-icon"></div>
-      <div class="plwd-toast-content"></div>
-    </div>
-  </body>
-</html>
