+++ test/core/mocks/mockPage.html
@@ -1,0 +1,48 @@
+<!-- Purpose: Mock Observer UI for testing dom manipulation -->
+<!doctype html>
+<html>
+  <head></head>
+  <body>
+    <div id="app">
+      <div class="very-big-div">
+        0540344327 4425770107 9974989020 2543431863 8106195438 5565504323 1825608663 7964780115 4577022857 0880546227
+        6167289712 5387678373 6696604931 0396451440 5526188732 1981989548 9352361347 4879892141 2295219449 3245682632
+        9387156775 2412205372 2190741310 0835318583 2913817091 7191308542 6730002552 1803605809 2896519591 2009405678
+        2196906889 7007107630 6325630088 5358448283 2675987213 0822101955 1903895604 8941299626 2816770155 7553516325
+        5229153539 5717387732 4281695351 7861891018 9785872180 7932711263 0622582090 0414521604 1887539191 1522049145
+        3304161927 4348928282 5961070285 6852113902 3406352336 1554696897 8785717171 1091125077 1643570309 4842045342
+        2789241460 6836272820 6152518578 9358210357 7850600762 6200727261 1366353838 6735738104 4666132012 2405225899
+        9214174715 7786184186 0250937063 0550161693 8184056432 4782236794 2177869917 1243214100 6792801366 9467844854
+        2579874532 9987087163 2721132531 6705082819 3967831568 6687845207 1388332315 3805072646 2851545503 8436777970
+        0640270262 9345911167 6813395194 6396810648 9771236761 4385255215 8312170298 2351600466 6061679436 1920965820
+      </div>
+    </div>
+    <div class="plwd-panel">
+      <div class="plwd-banner"></div>
+      <div class="plwd-input-box">
+        <input class="plwd-input" />
+        <div class="plwd-loader-box"></div>
+        <button class="plwd-cancel-btn"></button>
+        <button class="plwd-accept-btn"></button>
+      </div>
+      <div class="plwd-preview">
+        <p class="plwd-preview-title"></p>
+        <button class="plwd-dry-run-btn"></button>
+        <button class="plwd-clear-btn"></button>
+        <ul class="plwd-timeline">
+          <li class="plwd-timeline-item">
+            <span class="plwd-marker">✓</span>
+            input-value
+            <button class="plwd-delete-btn" data-index="0">
+              <svg class="plwd-trash-icon"></svg>
+            </button>
+          </li>
+        </ul>
+      </div>
+    </div>
+    <div class="plwd-toast">
+      <div class="plwd-toast-icon"></div>
+      <div class="plwd-toast-content"></div>
+    </div>
+  </body>
+</html>
