--- test/core/mocks/mockPageContent.html
@@ -1,27 +1,0 @@
-<!-- Purpose: Mock page content for testing page tools and assertion tools -->
-<!doctype html>
-<html>
-  <head>
-    <style>
-      .testClass {
-        color: red;
-      }
-    </style>
-  </head>
-  <body>
-    <noscript>No script</noscript>
-    <div id="app">
-      <div id="targetDiv">ID</div>
-      <div data-testid="testDiv">Test ID</div>
-      <div data-qa="testQA">Test QA</div>
-      <a href="https://test.url">Test URL</a>
-      <div class="testClass">Test Class</div>
-      <p style="color: red">Text</p>
-      <input type="text" />
-      <select>
-        <option value="1">Option 1</option>
-      </select>
-    </div>
-    <script type="text/javascript"></script>
-  </body>
-</html>
